Clean up  migration code.
